Move notificationsToIgnore to constants.
